青天井対応

+++ Assets/Package/Runtime/RoundResult.cs
@@ -7,7 +7,7 @@ namespace TSKT.Mahjongs
     public class RoundResult
     {
         readonly public GameResult gameResult;
-        public Dictionary<Player, int> scoreDiffs;
+        public Dictionary<Player, double> scoreDiffs;
         public Dictionary<Player, ExhausiveDrawType> states;
 
         public RoundResult(GameResult gameResult)
